This commit does all of the deconvolution.
The wire charges are deconvoluted to remove the shaping from the
electronics so that the shape looks like the time the charge arrives at the
wire (or passes the wire).  The induction planes are corrected so that
instead of an induction signal, it records current passing the wire.

The next step is to find the peaks in the ADC signals and translate that
into times and charges.