Improve the baseline (i.e. background) subtraction for deconvolved signals.
The baseline for deconvolved signals wanders around quite a bit,
particularly for the induction wires.  This is because of how the wire
noise interacts with the electronics response.  It's expected, and not much
of a problem except the peak finding algorithm has to deal with it.  The
background subtraction on the TSpectrum::SearchHighRes doesn't handle
situations where the peak width can vary widely, so the automatic
background handling has been turned off, and the baseline subtraction is
TPulseDeconvolution has been improved.