Change default for saving deconvoluted digits
The calibration has the ability to save the result of the channel
calibration and deconvolution for the TPC wires.  It's a great debugging
tool since shows what the hit finder is doing, but it increases a typical
file size almost x10.   The default for CLUSTERCALIB.exe is changed to NOT
save the calibrated ADC values, and only save the hits that have been
found.