Improve shape and normalization of the noise filter.
The noise filter is suppose to be based on the power in the (unknown)
signal.  The previous version used a hypothesis for the signal based on the
electronics response convoluted with the wire response, but this filtered
to much at the low frequencies for the induction wires.  The new hypthesis
is the simply the electronics response.  This also fixes the normalization
of the filter so that the maximum filter value is 1.0 (i.e. no filtering).