Improve the noise filter.
This bites the bullet and does a much better (i.e. slower) estimate of the signal power in the measured power spectrum.  The hypothesis is that the measured spectrum is a sum of the signal power, Gaussian noise, and an external "fixed" frequency power source.  The fixed frequency source introduces power at multiple, discrete, but unknown, frequencies.