Always cache the TCalibPulseDigits into the TEvent.
Previously, TCalibPulseDigits were only added to the TEvent if they were
going to be saved (i.e. with the CLUSTERCALIB.exe -O pulse option), but
that prevented looking all channels during calibration.  This changes the
logic so that TCalibPulseDigits are always added to the event, but are
added as a temporary object if they aren't going to be saved to the output
file.  This just implements saving the digits as temporaries, but doesn't
refactor to use the capability.