Make removing the correlated pedestal the default.
Removing the correlated pedestal needs further study, but it makes a
dramatic qualitative difference in the number of bogus hits that are
reconstructed, and seems to do a better job at estimating the charge in the
hits that are found (based on how much sample to sample fluctuations are
seen for a MIP).  Based on that, the correlated pedestal subtraction is now
the default.  It can be turned off from the command line.