Remove the electron lifetime correction from clusterCalib.
The electron lifetime correction was always out of place in clusterCalib
because it depended on knowing when an event happened.  Even within a beam
experiment, there can be multiple interactions at multiple locations, so
assigning the drift time is a reconstruction problem, not a calibration
problem.  This needs to be handled outside of clusterCalib inside the
reconstruction.