Improvements in the hit parameter error estimations.
This makes major changes in the way the uncertainty on the charge and time
are calculated when hits are created.  The charge uncertainty now includes
terms for the uncertainty in the baseline.  This is based on the variation
of the baseline in a limited range before and after the hit.  The hit time
is based on the peak position of the charge, and the curvature at the
peak, however, when the mean time is distant from the peak, the hit time
migrates to the mean time.  The hit uncertainty is calculated to always
include both the mean and the peak time of a charge distribution.